update base Fragment
