+++ app/src/main/java/com/catelt/mome/core/BaseFragment.kt
@@ -2,13 +2,20 @@ package com.catelt.mome.core
 
 import android.os.Bundle
 import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
+import androidx.viewbinding.ViewBinding
 
 typealias Inflate<T> = (LayoutInflater, ViewGroup?, Boolean) -> T
 
-abstract class BaseFragment : Fragment() {
+abstract class BaseFragment<VBinding : ViewBinding>(
+    private val inflate: Inflate<VBinding>
+): Fragment() {
+    protected lateinit var binding: VBinding
+
+    open val viewModel: BaseViewModel? get() = null
 
     open fun setUpViews() {}
 
@@ -21,5 +28,27 @@ abstract class BaseFragment : Fragment() {
 
     fun toast(message: String) {
         Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        arguments?.run {
+            setUpArgument(this)
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setUpAdapter()
+        setUpViews()
     }
 }
