--- app/src/main/java/com/catelt/mome/core/BaseFragmentNoViewModel.kt
@@ -1,37 +1,0 @@
-package com.catelt.mome.core
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.viewbinding.ViewBinding
-
-
-abstract class BaseFragmentNoViewModel<VBinding : ViewBinding>(
-    private val inflate: Inflate<VBinding>
-) : BaseFragment() {
-
-    lateinit var binding: VBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.run {
-            setUpArgument(this)
-        }
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        setUpAdapter()
-        setUpViews()
-    }
-}
